UI Images - pilot meters, gunner ammo
Added image elements for potential UI style - altitude meter, decided to use speed graphic as health meter instead - they overlay the current sliders for intended effect. Bullets on gunners are now on AmmoMonitor prefab - not animated/coded yet. the original ammo counter remains on there but it is intended to get a "Reload" text appear with a pie-circle timer graphic that fills in for the time that it takes to reload.