Dissolve Shader
See GraphicsTesting scene for how it looks:
Controls: DissolveDensity is the main 0-1f range that determines the state 0: solid, 1: fully dissolved. Animate this property for the effect.
maincolor and edge should should be self-explanatory
Dissolve Scale - higher it is, 'smaller' the noise is therefore best to adjust for giant objects
BurnScale - keep negative in small amount, this determines the side of the burning edge